remove warning message on clean callpickup member
warning:
/usr/lib/python2.7/dist-packages/sqlalchemy/sql/default_comparator.py:153:
SAWarning: The IN-predicate on "pickupmember.memberid" was invoked with
an empty sequence. This results in a contradiction, which nonetheless
can be expensive to evaluate.  Consider alternative strategies for
improved performance.  'strategies for improved performance.' % expr)
Updating database done

+++ alembic/versions/378c46e8c6fe_remove_pickupmember_invalid.py
@@ -40,24 +40,35 @@ def upgrade():
     queue_ids = [r.id for r in conn.execute(sql.select([queuefeatures.c.id])).fetchall()]
     user_ids = [r.id for r in conn.execute(sql.select([userfeatures.c.id])).fetchall()]
 
-    op.execute(
-        pickupmember
-        .delete()
-        .where(sql.or_(
+    filters = []
+    if group_ids:
+        filters.append(
             sql.and_(
                 pickupmember.c.membertype == 'group',
                 sql.not_(pickupmember.c.memberid.in_(group_ids)),
-            ),
+            )
+        )
+    if queue_ids:
+        filters.append(
             sql.and_(
                 pickupmember.c.membertype == 'queue',
                 sql.not_(pickupmember.c.memberid.in_(queue_ids)),
             ),
+        )
+    if user_ids:
+        filters.append(
             sql.and_(
                 pickupmember.c.membertype == 'user',
                 sql.not_(pickupmember.c.memberid.in_(user_ids)),
             ),
-        ))
-    )
+        )
+
+    if filters:
+        op.execute(
+            pickupmember
+            .delete()
+            .where(sql.or_(*filters))
+        )
 
 
 def downgrade():
